+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix checks using stale army indices after a player leaves
 - Fix defeated allies continuing to issue situation reports
 - Fix situation reports being duplicated when you have multiple AI allies
+- Fix allies with multiple commanders not reporting their landing location
 
 ## v1.2.1 - 2023-08-23