+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix situation reports being duplicated when you have multiple AI allies
 - Fix allies with multiple commanders not reporting their landing location
 - Fix an orbital fabber alone triggering a colonisation message
+- Fix Bugs gas giant colonisation never being reported
 
 ## v1.2.1 - 2023-08-23